+++ src/toisto/ui/diff.py
@@ -2,7 +2,7 @@
 
 from difflib import SequenceMatcher
 
-from .dictionary import linkify
+from .dictionary import linkified
 from .style import DELETED, INSERTED
 
 
@@ -25,7 +25,7 @@ def colored_diff(old_text: str, new_text: str, min_ratio_for_diff: float = 0.6) 
     """Return a colored string showing the diffs between old and new text."""
     matcher = SequenceMatcher(a=old_text.lower(), b=new_text.lower())
     if matcher.ratio() < min_ratio_for_diff:
-        return inserted(linkify(new_text))
+        return inserted(linkified(new_text))
     result = ""
     for operator, old_start, old_end, new_start, new_end in matcher.get_opcodes():
         old_fragment, new_fragment = (old_text[old_start:old_end], new_text[new_start:new_end])